Feature/WorldStartGeneration

+++ Plugins/CustomAssetsEditor/Source/CustomAssetsEditorRuntime/Public/BiomeAsset/Assets/BiomeAsset.h
@@ -29,19 +29,19 @@ public:
 	UFUNCTION(BlueprintCallable, Category="Utility|Search")
 	TArray<int32> GetRegisteredIDs() const noexcept;
 
-	TArray<TMap<int32, float>*> GetWeightMapsByID(int32 SearchID) const;
-	TMap<int32, float>* GetWeightMapByID(int32 SearchID, int32 Index) const;
-	TMap<int32, float>* GetWeightMapByType(const TSubclassOf<AActor>& SearchType, int32 Index) const;
+	TArray<TMap<int32, double>*> GetWeightMapsByID(int32 SearchID) const;
+	TMap<int32, double>* GetWeightMapByID(int32 SearchID, int32 Index) const;
+	TMap<int32, double>* GetWeightMapByType(const TSubclassOf<AActor>& SearchType, int32 Index) const;
 
 	UFUNCTION(BlueprintCallable, Category="Utility|Weight")
-	TMap<int32, float> GetTopWeightMapByID(const int32 SearchID) const noexcept {
+	TMap<int32, double> GetTopWeightMapByID(const int32 SearchID) const noexcept {
 		const auto Map = GetWeightMapByID(SearchID, 0);
 		if (Map == nullptr)
 			return {};
 		return *Map;
 	}
 	UFUNCTION(BlueprintCallable, Category="Utility|Weight")
-	TMap<int32, float> GetRightWeightMapByID(const int32 SearchID) const noexcept
+	TMap<int32, double> GetRightWeightMapByID(const int32 SearchID) const noexcept
 	{
 		const auto Map = GetWeightMapByID(SearchID, 1);
 		if (Map == nullptr)
@@ -49,7 +49,7 @@ public:
 		return *Map;
 	}
 	UFUNCTION(BlueprintCallable, Category="Utility|Weight")
-	TMap<int32, float> GetBottomWeightMapByID(const int32 SearchID) const noexcept
+	TMap<int32, double> GetBottomWeightMapByID(const int32 SearchID) const noexcept
 	{
 		const auto Map = GetWeightMapByID(SearchID, 2);
 		if (Map == nullptr)
@@ -57,7 +57,7 @@ public:
 		return *Map;
 	}
 	UFUNCTION(BlueprintCallable, Category="Utility|Weight")
-	TMap<int32, float> GetLeftWeightMapByID(const int32 SearchID) const noexcept
+	TMap<int32, double> GetLeftWeightMapByID(const int32 SearchID) const noexcept
 	{
 		const auto Map = GetWeightMapByID(SearchID, 3);
 		if (Map == nullptr)
@@ -66,7 +66,7 @@ public:
 	}
 	
 	UFUNCTION(BlueprintCallable, Category="Utility|Weight")
-	TMap<int32, float> GetTopWeightMapByType(const TSubclassOf<AActor> SearchType) const noexcept
+	TMap<int32, double> GetTopWeightMapByType(const TSubclassOf<AActor> SearchType) const noexcept
 	{
 		const auto Map = GetWeightMapByType(SearchType, 0);
 		if (Map == nullptr)
@@ -74,7 +74,7 @@ public:
 		return *Map;
 	}
 	UFUNCTION(BlueprintCallable, Category="Utility|Weight")
-	TMap<int32, float> GetRightWeightMapByType(const TSubclassOf<AActor> SearchType) const noexcept
+	TMap<int32, double> GetRightWeightMapByType(const TSubclassOf<AActor> SearchType) const noexcept
 	{
 		const auto Map = GetWeightMapByType(SearchType, 1);
 		if (Map == nullptr)
@@ -82,7 +82,7 @@ public:
 		return *Map;
 	}
 	UFUNCTION(BlueprintCallable, Category="Utility|Weight")
-	TMap<int32, float> GetBottomWeightMapByType(const TSubclassOf<AActor> SearchType) const noexcept
+	TMap<int32, double> GetBottomWeightMapByType(const TSubclassOf<AActor> SearchType) const noexcept
 	{
 		const auto Map = GetWeightMapByType(SearchType, 2);
 		if (Map == nullptr)
@@ -90,7 +90,7 @@ public:
 		return *Map;
 	}
 	UFUNCTION(BlueprintCallable, Category="Utility|Weight")
-	TMap<int32, float> GetLeftWeightMapByType(const TSubclassOf<AActor> SearchType) const noexcept {
+	TMap<int32, double> GetLeftWeightMapByType(const TSubclassOf<AActor> SearchType) const noexcept {
 		const auto Map = GetWeightMapByType(SearchType, 3);
 		if (Map == nullptr)
 			return {};
